refactor(agents): extract OpenCode SQLite source
